feat(1585): removed kms from receipt, as we do not need operator for receipt query
